+++ src/gort/devices/telescope.py
@@ -16,8 +16,9 @@ from typing import TYPE_CHECKING, ClassVar
 
 import numpy
 
+from gort.enums import Event
 from gort.exceptions import ErrorCode, GortTelescopeError
-from gort.gort import GortClient, GortDevice, GortDeviceSet
+from gort.gort import Gort, GortClient, GortDevice, GortDeviceSet
 from gort.tools import angular_separation, kubernetes_restart_deployment
 
 
@@ -44,12 +45,39 @@ class MoTanDevice(GortDevice):
         self.timeouts: defaultdict[str, float | None]
         self.timeouts = defaultdict(lambda: None, timeouts)
 
+        self.telescope = self.name.split(".")[0]
+
+        self.device: str
+        if "fibsel" in self.name:
+            self.device = "fibsel"
+        elif "foc" in self.name:
+            self.device = "focuser"
+        elif "km" in self.name:
+            self.device = "kmirror"
+        else:
+            raise GortTelescopeError(f"Invalid device type {self.name!r}.")
+
+    async def status(self):
+        """Returns the status of the device."""
+
+        return await self.run_command("status")
+
     async def is_reachable(self):
         """Is the device reachable?"""
 
         is_reachable = await self.run_command("isReachable")
 
         return bool(is_reachable.get("Reachable"))
+
+    async def check_reachable(self):
+        """Checks if the device is reacheable or issues an error."""
+
+        if not (await self.is_reachable()):
+            raise GortTelescopeError(
+                f"MoTAN device {self.name!r} is not reachable.",
+                error_code=ErrorCode.MOTAN_DEVICE_NOT_REACHABLE,
+                payload={"telescope": self.telescope, "device": self.device},
+            )
 
     async def is_moving(self):
         """Is the device moving."""
@@ -64,7 +92,23 @@ class MoTanDevice(GortDevice):
         if isinstance(self.SLEW_DELAY, (float, int)):
             await asyncio.sleep(self.SLEW_DELAY)
         else:
-            await asyncio.sleep(self.SLEW_DELAY[self.name.split(".")[0]])
+            await asyncio.sleep(self.SLEW_DELAY[self.telescope])
+
+    async def stop(self):
+        """Stop the K-mirror movement."""
+
+        if await self.is_moving():
+            self.write_to_log("Stopping slew.")
+
+            await self.run_command("slewStop", timeout=self.timeouts["slewStop"])
+            await self.run_command("stop", timeout=self.timeouts["slewStop"])
+
+            if await self.is_moving():
+                raise GortTelescopeError(
+                    f"Failed to stop the {self.name}.",
+                    error_code=ErrorCode.MOTAN_DEVICE_CANNOT_STOP,
+                    payload={"telescope": self.telescope, "device": self.device},
+                )
 
     async def run_command(
         self,
@@ -86,19 +130,13 @@ class KMirror(MoTanDevice):
 
     SLEW_DELAY = 0
 
-    async def status(self):
-        """Returns the status of the k-mirror."""
-
-        return await self.run_command("status")
-
     async def home(self):
         """Homes the k-mirror."""
 
-        if not (await self.is_reachable()):
-            raise GortTelescopeError("Device is not reachable.")
+        await self.check_reachable()
 
         await self.slew_delay()
-        await self.run_command("slewStop", timeout=self.timeouts["slewStop"])
+        await self.stop()
 
         self.write_to_log("Homing k-mirror.", level="info")
         await self.run_command("moveToHome", timeout=self.timeouts["moveToHome"])
@@ -106,11 +144,6 @@ class KMirror(MoTanDevice):
 
     async def park(self):
         """Park the k-mirror at 90 degrees."""
-
-        if not (await self.is_reachable()):
-            raise GortTelescopeError("Device is not reachable.")
-
-        await self.slew_delay()
 
         await self.move(90)
 
@@ -124,17 +157,12 @@ class KMirror(MoTanDevice):
 
         """
 
-        if not (await self.is_reachable()):
-            raise GortTelescopeError("Device is not reachable.")
+        await self.check_reachable()
 
         await self.slew_delay()
+        await self.stop()
 
         self.write_to_log(f"Moving k-mirror to {degs:.3f} degrees.", level="info")
-
-        self.write_to_log("Stopping slew.")
-        await self.run_command("slewStop", timeout=self.timeouts["slewStop"])
-
-        self.write_to_log("Moving k-mirror to absolute position.")
         await self.run_command(
             "moveAbsolute",
             degs,
@@ -168,8 +196,7 @@ class KMirror(MoTanDevice):
 
         """
 
-        if not (await self.is_reachable()):
-            raise GortTelescopeError("Device is not reachable.")
+        await self.check_reachable()
 
         await self.slew_delay()
 
@@ -191,6 +218,7 @@ class KMirror(MoTanDevice):
         if abs(stop_degs_before) > 0:
             self.write_to_log(f"Using stop_degs_before={stop_degs_before}.")
 
+        await self.stop()
         await self.run_command(
             "slewStart",
             ra / 15.0,
@@ -207,11 +235,6 @@ class Focuser(MoTanDevice):
 
     SLEW_DELAY = 0
 
-    async def status(self):
-        """Returns the status of the focuser."""
-
-        return await self.run_command("status")
-
     async def home(self):
         """Homes the focuser.
 
@@ -222,14 +245,14 @@ class Focuser(MoTanDevice):
 
         """
 
-        if not (await self.is_reachable()):
-            raise GortTelescopeError("Device is not reachable.")
+        await self.check_reachable()
+
+        await self.slew_delay()
+        await self.stop()
 
         # Store current position to restore it later.
         status = await self.status()
         current_position = status.get("Position")
-
-        await self.slew_delay()
 
         self.write_to_log("Homing focuser.", level="info")
         await self.run_command("moveToHome", timeout=self.timeouts["moveToHome"])
@@ -242,10 +265,10 @@ class Focuser(MoTanDevice):
     async def move(self, dts: float):
         """Move the focuser to a position in DT."""
 
-        if not (await self.is_reachable()):
-            raise GortTelescopeError("Device is not reachable.")
+        await self.check_reachable()
 
         await self.slew_delay()
+        await self.stop()
 
         self.write_to_log(f"Moving focuser to {dts:.3f} DT.", level="info")
         await self.run_command(
@@ -271,18 +294,13 @@ class FibSel(MoTanDevice):
 
         self.__last_homing: float = 0
 
-    async def status(self):
-        """Returns the status of the fibre selector."""
-
-        return await self.run_command("status")
-
     async def home(self):
         """Homes the fibre selector."""
 
-        if not (await self.is_reachable()):
-            raise GortTelescopeError("Device is not reachable.")
+        await self.check_reachable()
 
         await self.slew_delay()
+        await self.stop()
 
         self.write_to_log("Homing fibsel.", level="info")
         await self.run_command("moveToHome", timeout=self.timeouts["moveToHome"])
@@ -319,13 +337,10 @@ class FibSel(MoTanDevice):
 
         """
 
-        if not (await self.is_reachable()):
-            raise GortTelescopeError("Device is not reachable.")
+        await self.check_reachable()
 
         if rehome:
             await self.home()
-        else:
-            await self._check_home()
 
         if isinstance(position, str):
             mask_positions = self.gort.config["telescopes"]["mask_positions"]
@@ -342,29 +357,46 @@ class FibSel(MoTanDevice):
             steps = position
             self.write_to_log(f"Moving mask to {steps} DT.", level="info")
 
-        await self.slew_delay()
-        await self.run_command(
-            "moveAbsolute",
-            steps,
-            timeout=self.timeouts["moveAbsolute"],
-        )
+        await self._move("moveAbsolute", steps)
 
     async def move_relative(self, steps: float):
         """Move the mask a number of motor steps relative to the current position."""
 
-        if not (await self.is_reachable()):
-            raise GortTelescopeError("Device is not reachable.")
+        await self.check_reachable()
 
         self.write_to_log(f"Moving fibre mask {steps} steps.")
+        await self._move("moveRelative", steps)
 
-        await self.slew_delay()
-        await self._check_home()
+    async def _move(self, command: str, steps: float, allow_rehoming: bool = True):
+        """Moves the fibre selector. If the move fails, tries rehoming."""
 
-        await self.run_command(
-            "moveRelative",
-            steps,
-            timeout=self.timeouts["moveRelative"],
-        )
+        try:
+            await self._check_home()
+
+            await self.slew_delay()
+            await self.stop()
+
+            await self.run_command(command, steps, timeout=self.timeouts[command])
+        except Exception as err:
+            if allow_rehoming:
+                self.write_to_log(
+                    f"Failed to move fibsel with error: {err}. Rehoming and retrying.",
+                    "warning",
+                )
+
+                if isinstance(self.gort, Gort):
+                    # Notify this event so that it can be taken into account, for
+                    # example to add a flag if this happens during the standards
+                    # iteration loop.
+                    await self.gort.notify_event(
+                        Event.UNEXPECTED_FIBSEL_REHOME,
+                        payload={"time": time()},
+                    )
+
+                await self.home()
+                await self._move(command, steps, allow_rehoming=False)
+            else:
+                raise
 
 
 class Telescope(GortDevice):
@@ -550,11 +582,6 @@ class Telescope(GortDevice):
 
         await self.initialise()
 
-        kmirror_task: asyncio.Task | None = None
-        if kmirror and self.km:
-            self.write_to_log("Parking k-mirror.", level="info")
-            kmirror_task = asyncio.create_task(self.km.park())
-
         if use_pw_park:
             self.write_to_log("Parking telescope to PW default position.", level="info")
             await self.pwi.commands.park()
@@ -584,8 +611,9 @@ class Telescope(GortDevice):
             self.write_to_log("Disabling telescope.")
             await self.pwi.commands.setEnabled(False)
 
-        if kmirror_task:
-            await kmirror_task
+        if kmirror and self.km:
+            self.write_to_log("Parking k-mirror.", level="info")
+            await self.km.park()
 
         self.is_homed = False