+++ src/gort/exceptions.py
@@ -8,13 +8,11 @@
 
 from __future__ import annotations
 
-import asyncio
 import inspect
 
 from typing import TYPE_CHECKING, ClassVar
 
-from gort.enums import ErrorCode, Event
-from gort.pubsub import notify_event
+from gort.enums import ErrorCode
 
 
 if TYPE_CHECKING:
@@ -38,7 +36,6 @@ class GortError(Exception):
         message: str | None = None,
         error_code: int | ErrorCode | None = None,
         payload: dict = {},
-        emit_event: bool | None = None,
     ):
         try:
             self.error_code = ErrorCode(error_code or self.DEFAULT_ERROR_CODE)
@@ -46,17 +43,7 @@ class GortError(Exception):
             self.error_code = ErrorCode.UNKNOWN_ERROR
             error_code = self.error_code.value
 
-        if emit_event is None:
-            emit_event = self.EMIT_EVENT
-
         self.payload = payload
-
-        if emit_event:
-            event_payload = self.payload.copy()
-            event_payload["error"] = message or ""
-            event_payload["error_code"] = self.error_code.value
-
-            asyncio.create_task(notify_event(Event.ERROR, payload=event_payload))
 
         prefix = f"Error {self.error_code.value} ({self.error_code.name})"
         if message is not None and message != "":
@@ -147,6 +134,8 @@ class GortDeviceError(GortError):
         self,
         message: str | None = None,
         error_code: int | ErrorCode | None = None,
+        payload: dict = {},
+        emit_event: bool | None = None,
     ) -> None:
         from gort.gort import GortDevice
 
@@ -160,7 +149,7 @@ class GortDeviceError(GortError):
                 if issubclass(obj.__class__, GortDevice) and name is not None:
                     message = f"({name}) {message}"
 
-        super().__init__(message, error_code=error_code)
+        super().__init__(message, error_code=error_code, payload=payload)
 
 
 class GortEnclosureError(GortDeviceError):