+++ src/gort/gort.py
@@ -16,6 +16,8 @@ import subprocess
 import sys
 import uuid
 from copy import deepcopy
+from functools import partial
+from types import TracebackType
 
 from typing import (
     TYPE_CHECKING,
@@ -203,48 +205,111 @@ class GortClient(AMQPClient):
 
         return log
 
-    def _setup_exception_hooks(self, log: SDSSLogger, use_rich_output: bool = True):
+    async def notify_event(self, event: Event, payload: dict[str, Any] = {}):
+        """Emits an event notification."""
+
+        try:
+            await notify_event(event, payload=payload)
+        except TypeError as err:
+            if "is not JSON serializable" in str(err):
+                self.log.error(
+                    f"Failed to notify event {event.name}: payload is not"
+                    "serialisable. The event will be emitted without payload."
+                )
+                await notify_event(event)
+
+    def exception_handler(
+        self,
+        log: SDSSLogger,
+        exc_type: type[BaseException] | None,
+        exc_value: BaseException | None,
+        exc_traceback: TracebackType | None,
+    ):
+        """A custom exception handler that logs exceptions to file."""
+
+        log.handle_exceptions(exc_type, exc_value, exc_traceback)
+
+        if exc_value and isinstance(exc_value, GortError):
+            event_payload = exc_value.payload.copy()
+            event_payload["error"] = exc_value.args[0] or ""
+            event_payload["error_code"] = exc_value.error_code.value
+
+            try:
+                loop = asyncio.get_running_loop()
+            except RuntimeError:  # No running event loop, mostly for IPython.
+                log.warning("Cannot emit error event: no running event loop.")
+            else:
+                loop.create_task(self.notify_event(Event.ERROR, payload=event_payload))
+
+    def asyncio_exception_handler(self, loop, context):
+        """Handle an uncaught asyncio exception and reports it."""
+
+        if exception := context.get("exception", None):
+            try:
+                raise exception
+            except Exception:
+                exc_type, exc_value, exc_tb = sys.exc_info()
+                self.exception_handler(self.log, exc_type, exc_value, exc_tb)
+        else:
+            loop.default_exception_handler(context)
+
+    def _setup_exception_hooks(
+        self,
+        log: SDSSLogger | None = None,
+        use_rich_output: bool = True,
+    ):
         """Setup various hooks for exception handling."""
+
+        log = log or self.log
 
         def custom__showtraceback_closure(default__showtraceback):
             def _showtraceback(*args, **kwargs):
                 assert IPYTHON
 
                 exc_tuple = IPYTHON._get_exc_info()
-                log.handle_exceptions(*exc_tuple)
-
-                default__showtraceback(*args, **kwargs)
+                self.exception_handler(log, *exc_tuple)
 
             if IPYTHON:
                 IPYTHON._showtraceback = _showtraceback
 
         if use_rich_output:
-            traceback.install(console=self._console)
-            pretty.install(console=self._console)
+            if not IPYTHON:
+                traceback.install(console=self._console)
+                pretty.install(console=self._console)
 
             # traceback.install() overrides the excepthook, which means that
             # tracebacks are not logged to file anymore. Restore that.
-            sys.excepthook = log.handle_exceptions
+            sys.excepthook = partial(self.exception_handler, log)
 
         else:
             # Make sure we are not using rich tracebacks anymore, in
             # case we installed them at some point.
             # See https://github.com/Textualize/rich/pull/2972/files
 
-            sys.excepthook = log.handle_exceptions
+            sys.excepthook = partial(self.exception_handler, log)
 
             if IPYTHON:
                 IPYTHON._showtraceback = IPYTHON_DEFAULT_HOOKS[0]
                 IPYTHON.showtraceback = IPYTHON_DEFAULT_HOOKS[1]
                 IPYTHON.showsyntaxerror = IPYTHON_DEFAULT_HOOKS[2]
 
-        # if IPYTHON:
-        #     # One more override. We want that any exceptions raised in IPython
-        #     # also gets logged to file, but IPython overrides excepthook completely
-        #     # so here we make a custom call to log.handle_exceptions() and then
-        #     # just let it do whatever it was its default (whether that means it
-        #     # was overridden by rich or not).
-        #     custom__showtraceback_closure(IPYTHON._showtraceback)
+        if IPYTHON:
+            # One more override. We want that any exceptions raised in IPython
+            # also gets logged to file, but IPython overrides excepthook completely
+            # so here we make a custom call to log.handle_exceptions() and then
+            # just let it do whatever it was its default (whether that means it
+            # was overridden by rich or not).
+            custom__showtraceback_closure(IPYTHON._showtraceback)
+
+    def _setup_async_exception_hooks(self):
+        """Sets up a custom async exception handler."""
+
+        try:
+            loop = asyncio.get_running_loop()
+        except RuntimeError:  # No running event loop, mostly for IPython.
+            pass
+        else:
+            loop.set_exception_handler(self.asyncio_exception_handler)
 
     def get_log_path(self):
         """Returns the path of the log file. :obj:`None` if not logging to file."""
@@ -264,6 +329,11 @@ class GortClient(AMQPClient):
         if not self.connected:
             async with self._connect_lock:
                 await self.start()
+
+        # Override the asyncio exception handler to catch errors in tasks.
+        # We do this after AMQPClient.start() because it sets its own exception
+        # handler.
+        self._setup_async_exception_hooks()
 
         override_overwatcher = getattr(self, "_override_overwatcher", None)
         override_envvar = os.environ.get("GORT_OVERRIDE_OVERWATCHER", "0")
@@ -662,19 +732,6 @@ class Gort(GortClient):
 
         if verbosity:
             self.set_verbosity(verbosity)
-
-    async def notify_event(self, event: Event, payload: dict[str, Any] = {}):
-        """Emits an event notification."""
-
-        try:
-            await notify_event(event, payload=payload)
-        except TypeError as err:
-            if "is not JSON serializable" in str(err):
-                self.log.error(
-                    f"Failed to notify event {event.name}: payload is not"
-                    "serialisable. The event will be emitted without payload."
-                )
-                await notify_event(event)
 
     @Retrier(max_attempts=3, delay=3)
     async def emergency_shutdown(self):